+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 - Scannerjobs planned for 23:45 are now able to be automatically run.
 
+- Exclusion rules can once again be chosen on the scanner forms in the UI.
+
 ## Version 3.21.4, 15th December 2023
 
 "Don't Crash, Bandicoot"